requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
requests
twilio
scikit-learn==1.2.2
numpy==1.23.5